Fix check 2 data types

+++ python-abc/task_01_duck_typing.py
@@ -29,11 +29,11 @@ class Circle(Shape):
 
     def area(self):
         """Return area of circle."""
-        return math.pi * (self.radius ** 2)
+        return float(math.pi * (self.radius ** 2))
 
     def perimeter(self):
         """Return perimeter of circle."""
-        return 2 * math.pi * self.radius
+        return float(2 * math.pi * self.radius)
 
 
 class Rectangle(Shape):
@@ -46,15 +46,14 @@ class Rectangle(Shape):
 
     def area(self):
         """Return area of rectangle."""
-        return self.width * self.height
+        return float(self.width * self.height)
 
     def perimeter(self):
         """Return perimeter of rectangle."""
-        return 2 * (self.width + self.height)
+        return float(2 * (self.width + self.height))
 
 
 def shape_info(shape):
     """Print area and perimeter using duck typing."""
     print("Area: {}".format(shape.area()))
     print("Perimeter: {}".format(shape.perimeter()))
-
